movidas funciones de grupos a modulo
